pyproject.toml
[tool.poetry]
name = "air-sdk"
version = "2.13.0"
description = "Python SDK for interacting with NVIDIA Air"
license = "MIT"
classifiers = ["Programming Language :: Python :: 3", "License :: OSI Approved :: MIT License", "Operating System :: OS Independent"]
homepage = "https://github.com/NVIDIA/air_sdk"
authors = ["NVIDIA Air <air-support@nvidia.com>"]
readme = "README.md"

[tool.poetry.urls]
"Homepage" = "https://github.com/NVIDIA/air_sdk/issues"
"Bug Tracker" = "https://github.com/NVIDIA/air_sdk/issues"

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.9.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
coverage = "^7.5.4"
faker = "^26.0.0"
mypy = "1.10.1"
pre-commit = "^2.21"
pytest = "^7.4.4"
requests-mock = "^1.12.1"
ruff = "0.5.3"

[tool.coverage.run]
omit = [
    "./tests/*",
    "__init__.py"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
